additional packages and widget file renaming
