--- app/src/main/res/layout/step_widget.xml
@@ -1,34 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/rl_step_widget_holder"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/bg_step_stack_item"
-    android:padding="@dimen/widget_margin">
-
-    <ImageView
-        android:id="@+id/img_step_cake_resource"
-        android:layout_width="@dimen/box_big"
-        android:layout_height="@dimen/box_big"
-        android:layout_below="@+id/step_recipe_name"
-        android:layout_centerHorizontal="true"
-        android:layout_margin="@dimen/margin_medium"
-        android:alpha="0.8"
-        android:contentDescription="@string/step_widget_preview"
-        tools:src="@drawable/cheesecake" />
-
-    <TextView
-        android:id="@+id/step_recipe_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:gravity="center_horizontal"
-        android:padding="8dp"
-        android:text="@string/appwidget_text"
-        android:textColor="@color/textPrimary"
-        android:textSize="24sp"
-        android:textStyle="bold|italic"
-        tools:text="Yellow Cake (step 5)" />
-
-</RelativeLayout>
