+++ app/src/main/res/xml/stack_widget_info.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:initialKeyguardLayout="@layout/step_widget"
-    android:initialLayout="@layout/step_widget"
+    android:initialKeyguardLayout="@layout/item_recipe_stack"
+    android:initialLayout="@layout/item_recipe_stack"
     android:minHeight="110dp"
     android:minWidth="180dp"
     android:previewImage="@drawable/recipe_stack_preview"
